fixin error in user seeder
